+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   FlawCVSS, FlawReference, Snippet and Tracker. (OSIDB-3466)
 - Moved metadata creation during tests to root level conftest and
   automatically set envs during VCR recording (OSIDB-3492)
+- Exclude component and version from Jira tracker updates (OSIDB-3677)
 
 ## [4.5.6] - 2024-11-08
 ### Fixed